refactor(eslint-plugin-i18n): import clearSpaceCharacters from shared directly
