+++ packages/eslint-plugin-i18n/rules/auto-generate-translation-message-id/utils/__tests__/clear-space-characters.spec.ts
@@ -1,4 +1,4 @@
-import {clearSpaceCharacters} from '../clear-space-characters';
+import {clearSpaceCharacters} from '../../../shared/clear-space-characters';
 
 describe('clearSpaceCharacters', () => {
     it('should remove newlines and leading spaces', () => {
