--- packages/eslint-plugin-i18n/rules/auto-generate-translation-message-id/utils/clear-space-characters.ts
@@ -1,1 +1,0 @@
-export {clearSpaceCharacters} from '../../shared/clear-space-characters';
